Add support for extension modules from PyPI
This adds support for installing binary wheels that are not named using
the "linux_roborio" platform tag, since this tag is not allowed on PyPI,
but "linux_armv7l" (the platform tag for the RoboRIO) is.

The changes here also easily allow for adding or changing platforms, if
either of these should ever become relevant.